split metadata to metadata
