+++ src/dailyphoto/new.py
@@ -5,6 +5,7 @@ from datetime import datetime
 from datetime import timedelta
 
 from . import config
+from .metadata import get_metadata_filename
 
 
 def new(
@@ -113,12 +114,12 @@ def new_image(
     shutil.move(old_image_path, new_image_path)
 
     # Try to move the metadata file
-    old_metadata_file = config.get_metadata_filename(
+    old_metadata_file = get_metadata_filename(
         config.UNUSED_METADATA,
         new_image,
     )
     if os.path.exists(old_metadata_file):
-        new_metadata_file = config.get_metadata_filename(
+        new_metadata_file = get_metadata_filename(
             config.METADATA_DIR,
             new_image,
         )
